[WIP] address requested changes
