+++ package.json
@@ -65,29 +65,30 @@
                     "type": "array",
                     "items": {
                         "type": "object",
-                        "title": "inner objects",
+                        "title": "Dependency module location",
                         "properties": {
                             "url": {
                                 "type": "string",
-                                "description": "Url of the git repo or pip index",
+                                "description": "Dependency module location url.",
                                 "order": 1
                             },
                             "type":{
                                 "type": "string",
                                 "enum": ["git", "package"],
-                                "enumDescriptions": [
-                                  "Url of the git repository containing a v1 module",
-                                  "Pip index url in which to look for v2 modules"
+                                "markdownEnumDescriptions": [
+                                  "URI of the parent location for a v1 module. (Organization url for github, Group url for gitlab, parent folder for a local repository...)",
+                                  "Pip index url in which to look for v2 modules. ⚠️ Be mindful of [dependency confusion attacks](https://www.websecuritylens.org/how-dependency-confusion-attack-works-and-how-to-prevent-it/) when setting this up."
                                 ],
+                                "description": "Type of this Dependency module location, one of: [\"git\", \"package\"].",
                                 "order": 2
                             }
-                        }
+                        },
+                        "markdownDescription": "Dependency module location."
                     },
                     "default": [
-                        {"url": "https://github.com/inmanta/", "type": "git"},
-                        {"url": "https://pypi.org/simple/", "type": "package"}
+                        {"url": "https://github.com/inmanta/", "type": "git"}
                     ],
-                    "markdownDescription": "List of git repositories and pip indexes to indicate the Inmanta extension where to look for modules when working on an individual module.\nWhen working on a project, this option is ignored and the Inmanta extension will look for modules in the repositories and pip indexes defined in the repo section of the [project.yaml](https://docs.inmanta.com/community/latest/reference/projectyml.html#project-yml) file.",
+                    "markdownDescription": "List of dependency module locations to indicate the Inmanta extension where to look for dependency modules when working on an individual module.\nWhen working on a project, this option is ignored and the Inmanta extension will look for modules in the repositories and pip indexes defined in the repo section of the [project.yaml](https://docs.inmanta.com/community/latest/reference/projectyml.html#project-yml) file.",
                     "scope": "resource"
                 }
             }
@@ -148,6 +149,14 @@
                         "description": "Install the language server to enjoy features like Syntax highlighting, Code navigation and many more. \n[Install Language Server](command:inmanta.installLS)",
                         "media": {
                             "markdown": "media/markdown_install_ls.md"
+                        }
+                    },
+                    {
+                        "id": "inmanta.setModuleDependencyLocations",
+                        "title": "Setup module dependency locations",
+                        "description": "Edit setting.\n[Go to setting](command:workbench.action.openSettings?%5B%22inmanta%3A%20Repos%22%5D)",
+                        "media": {
+                            "markdown": "media/markdown_set_module_dependency.md"
                         }
                     },
                     {
