+++ server/src/inmantals/server.py
@@ -318,7 +318,7 @@ class InmantaLSHandler(JsonRpcHandler):
 
         if init_options:
             self.compiler_venv_path = init_options.get(
-                "compilerVenv", os.path.join(os.path.abspath(urlparse(rootUri).path), ".env-ls-compiler")
+                "compilerVenv", os.path.join(os.path.abspath(urlparse(workspace_folder.uri).path), ".env-ls-compiler")
             )
             self.repos = init_options.get("repos", None)
             logger.debug("self.repos= %s", self.repos)
